scripts/agent_backup_20260308_120211/agents/support/farmcoach_agent_pkg/requirements.txt
pydantic>=1.8.0
pytest>=6.0.0
requests>=2.25.0
beautifulsoup4>=4.9.0
python-dotenv>=0.19.0
